+++ db/PostgreSQL/inventory-blank.sql
@@ -2155,6 +2155,10 @@ $$
     DECLARE _is_periodic                    boolean = inventory.is_periodic_inventory(_office_id);
     DECLARE _default_currency_code          national character varying(12);
 BEGIN
+    IF NOT finance.can_post_transaction(_login_id, _user_id, _office_id, _book_name, _value_date) THEN
+        return 0;
+    END IF;
+    
     CREATE TEMPORARY TABLE IF NOT EXISTS temp_stock_details
     (
         tran_type                       national character varying(2),
@@ -2339,7 +2343,7 @@ LANGUAGE plpgsql;
 -- ROW('Cr', '11MBA', 'Piece', 1)::inventory.adjustment_type
 -- ]
 -- );
--- 
+
 
 
 -->-->-- src/Frapid.Web/Areas/MixERP.Inventory/db/PostgreSQL/2.x/2.0/src/02.functions-and-logic/inventory.post_opening_inventory.sql --<--<--
@@ -2376,6 +2380,10 @@ $$
     DECLARE _tran_counter                   integer;
     DECLARE _transaction_code               text;
 BEGIN
+    IF NOT finance.can_post_transaction(_login_id, _user_id, _office_id, _book_name, _value_date) THEN
+        return 0;
+    END IF;
+
     DROP TABLE IF EXISTS temp_stock_details;
     
     CREATE TEMPORARY TABLE temp_stock_details
@@ -2493,6 +2501,10 @@ $$
     DECLARE _checkout_id                    bigint;
     DECLARE _book_name                      text='Inventory Transfer';
 BEGIN
+    IF NOT finance.can_post_transaction(_login_id, _user_id, _office_id, _book_name, _value_date) THEN
+        RETURN 0;
+    END IF;
+
     CREATE TEMPORARY TABLE IF NOT EXISTS temp_stock_details
     (
         tran_type       national character varying(2),
@@ -2770,6 +2782,50 @@ AND inventory.items.maintain_inventory;
 
 
 
+-->-->-- src/Frapid.Web/Areas/MixERP.Inventory/db/PostgreSQL/2.x/2.0/src/05.views/inventory.checkout_view.sql --<--<--
+DROP VIEW IF EXISTS inventory.checkout_view CASCADE;
+
+CREATE VIEW inventory.checkout_view
+AS
+SELECT
+        finance.transaction_master.transaction_master_id,
+        inventory.checkouts.checkout_id,
+        inventory.checkout_details.checkout_detail_id,
+        finance.transaction_master.book,
+        finance.transaction_master.transaction_counter,
+        finance.transaction_master.transaction_code,
+        finance.transaction_master.value_date,
+        finance.transaction_master.transaction_ts,
+        finance.transaction_master.login_id,
+        finance.transaction_master.user_id,
+        finance.transaction_master.office_id,
+        finance.transaction_master.cost_center_id,
+        finance.transaction_master.reference_number,
+        finance.transaction_master.statement_reference,
+        finance.transaction_master.last_verified_on,
+        finance.transaction_master.verified_by_user_id,
+        finance.transaction_master.verification_status_id,
+        finance.transaction_master.verification_reason,
+        inventory.checkout_details.transaction_type,
+        inventory.checkout_details.store_id,
+        inventory.checkout_details.item_id,
+        inventory.checkout_details.quantity,
+        inventory.checkout_details.unit_id,
+        inventory.checkout_details.base_quantity,
+        inventory.checkout_details.base_unit_id,
+        inventory.checkout_details.price,
+        inventory.checkout_details.discount,
+        inventory.checkout_details.shipping_charge,
+        inventory.checkout_details.price * inventory.checkout_details.quantity + inventory.checkout_details.discount AS amount
+FROM inventory.checkout_details
+INNER JOIN inventory.checkouts
+ON inventory.checkouts.checkout_id = inventory.checkout_details.checkout_id
+INNER JOIN finance.transaction_master
+ON finance.transaction_master.transaction_master_id = inventory.checkouts.transaction_master_id;
+
+
+
+
 -->-->-- src/Frapid.Web/Areas/MixERP.Inventory/db/PostgreSQL/2.x/2.0/src/05.views/inventory.item_view.sql --<--<--
 DROP VIEW IF EXISTS inventory.item_view;
 
@@ -2820,6 +2876,17 @@ ON finance.transaction_master.transaction_master_id = inventory.checkouts.transa
 AND finance.transaction_master.verification_status_id > 0;
 
 
+-->-->-- src/Frapid.Web/Areas/MixERP.Inventory/db/PostgreSQL/2.x/2.0/src/05.views/inventory.verified_checkout_view.sql --<--<--
+DROP MATERIALIZED VIEW IF EXISTS inventory.verified_checkout_view;
+
+CREATE MATERIALIZED VIEW inventory.verified_checkout_view
+AS
+SELECT * FROM inventory.checkout_view
+WHERE verification_status_id > 0;
+
+ALTER MATERIALIZED VIEW inventory.verified_checkout_view
+OWNER TO mix_erp;
+
 -->-->-- src/Frapid.Web/Areas/MixERP.Inventory/db/PostgreSQL/2.x/2.0/src/99.ownership.sql --<--<--
 DO
 $$
@@ -2835,6 +2902,25 @@ BEGIN
     AND tableowner <> 'frapid_db_user'
     LOOP
         EXECUTE 'ALTER TABLE '|| this.schemaname || '.' || this.tablename ||' OWNER TO frapid_db_user;';
+    END LOOP;
+END
+$$
+LANGUAGE plpgsql;
+
+DO
+$$
+    DECLARE this record;
+BEGIN
+    IF(CURRENT_USER = 'frapid_db_user') THEN
+        RETURN;
+    END IF;
+
+    FOR this IN 
+    SELECT oid::regclass::text as mat_view
+    FROM   pg_class
+    WHERE  relkind = 'm'
+    LOOP
+        EXECUTE 'ALTER TABLE '|| this.mat_view ||' OWNER TO frapid_db_user;';
     END LOOP;
 END
 $$
@@ -3010,3 +3096,5 @@ BEGIN
 END
 $$
 LANGUAGE plpgsql;
+
+
